gpu abi: i64 buffer size + sized SamplerDesc; evolution conventions; metal divergence fixes
- gpu.create_buffer carries an i64 size (the int cap topped out at 2GB and
  overflowed negative past it); GPUBackend::createBuffer widened to u64.
  exec_gpu.h mirrors; web hosts convert the BigInt.
- gpu.create_sampler takes a SIZED descriptor struct (first field = its own
  byte size; hosts read min(sent, known) and default the rest, so appending
  options later needs no new import name). SamplerDesc adds separate
  min/mag/mip filters, per-axis address modes, anisotropy, and lod clamps —
  and closes the mip-filter divergence (Metal hardcoded Linear while web
  followed the mag filter; both backends now honor the explicit field).
  Convenience Device::createSampler(filter, address) wrapper unchanged for
  effects.
- module_api.h documents the evolution conventions per call class: name-keyed
  effect exports (new hook = new name), flat-args + new-import-name for
  hot-path calls, sized descriptor structs for setup calls whose options
  accumulate, and hard-layout binary blobs pinned to NANO_ABI_VERSION.
- Metal divergence fixes riding along: getTextureFormat reports sRGB textures
  as code 7 (was misreported as plain RGBA8, web disagreed); a compute MSL
  kernel missing the nano_threadgroup hint now logs LOUDLY before the 8x8
  fallback (a non-8x8 kernel silently ran wrong on Metal only).

All bundles + executor rebuilt; 434/434 ctest, 1006/1006 vitest, e2e sweep
(blur/pipeline/platform) green.

+++ native/src/sketch/exec_gpu.h
@@ -50,7 +50,7 @@ EXEC_GPU_IMPORT("create_shader_module") int32_t gpu_create_shader_module(const c
 EXEC_GPU_IMPORT("create_compute_pso")   int32_t gpu_create_compute_pso(int32_t shader, const char* entry, int32_t entry_len);
 
 // Small uniform buffers (the wet/dry opacity blend pass).
-EXEC_GPU_IMPORT("create_buffer") int32_t gpu_create_buffer(int32_t size, int32_t usage);
+EXEC_GPU_IMPORT("create_buffer") int32_t gpu_create_buffer(int64_t size, int32_t usage);
 EXEC_GPU_IMPORT("write_buffer")  void    gpu_write_buffer(int32_t buf, int32_t offset, const void* data, int32_t data_len);
 
 // Compute pass dispatch.
